ai-fashion-platform/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
email-validator==2.2.0

# --- LLM SDKs (both optional; only the one you use needs to be installed)
anthropic==0.39.0
openai==1.54.3

# --- .env loading
python-dotenv==1.0.1
